# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stager API
  description: |
    API for Stager - AI-Powered Virtual Staging Platform for Real Estate.

    ## Authentication
    All endpoints (except webhooks) require authentication via Supabase Auth session cookies.
    Unauthenticated requests return `401 Unauthorized`.

    ## Rate Limiting
    Resource-intensive endpoints are rate-limited:
    - **Staging:** 20 requests/minute
    - **Email:** 10 requests/minute
    - **Auth:** 5 requests/minute

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Responses
    All errors follow a consistent format with actionable suggestions:
    ```json
    {
      "error": "Human-readable message",
      "code": "ERROR_CODE",
      "suggestion": "Actionable suggestion for the user",
      "action": { "label": "Button label", "href": "/path" },
      "details": { "field": "value" }
    }
    ```

    ## Credits
    - Staging costs 1 credit per image
    - First 2 remixes per original image are free
    - Additional remixes cost 1 credit each
  version: 1.136.0
  contact:
    name: Stager Support
servers:
  - url: https://stager-tau.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
paths:
...
  /team/invite/accept:
    get:
      tags: [Team]
      summary: Validate invitation token
      description: Check if an invitation token is valid and not expired.
      operationId: validateInvitation
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Invitation token from email link
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  invitation:
                    $ref: '#/components/schemas/InvitationDetails'
        '400':
          description: Invalid or expired token

    post:
      tags: [Team]
      summary: Accept invitation
      description: Accept an invitation and join the organization. User must be authenticated.
      operationId: acceptInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Invitation token
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  organizationId:
                    type: string
                    format: uuid
        '400':
          description: Invalid or expired token
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /team/invitations/{id}:
    delete:
      tags: [Team]
      summary: Revoke invitation
      description: Revoke a pending invitation. The token will no longer be valid.
      operationId: revokeInvitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invitation revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - UNAUTHORIZED
        - SESSION_EXPIRED
        - FORBIDDEN
        - NOT_OWNER
        - INSUFFICIENT_PERMISSIONS
        - INSUFFICIENT_CREDITS
        - PAYMENT_REQUIRED
        - RATE_LIMITED
        - VALIDATION_ERROR
        - INVALID_INPUT
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - PROVIDER_ERROR
        - EMAIL_ERROR
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
      description: Standardized error codes

    # ---- Request Schemas ----
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCode'
        suggestion:
          type: string
          description: Actionable suggestion for the user
        action:
          type: object
          description: Optional CTA button
          properties:
            label:
              type: string
            href:
              type: string
        details:
          type: object
          additionalProperties: true
          description: Additional context about the error
...
    InvitationDetails:
      type: object
      description: Full invitation details for accept page
      properties:
        email:
          type: string
          format: email
        organizationName:
          type: string
        inviterName:
          type: string
        initialCredits:
          type: integer
        expiresAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "You must be logged in to perform this action."
            code: "UNAUTHORIZED"
            suggestion: "Please log in or create an account to continue."
            action:
              label: "Log In"
              href: "/login"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "This action is restricted to organization owners."
            code: "INSUFFICIENT_PERMISSIONS"
            suggestion: "Contact your organization owner for assistance."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "The requested resource could not be found."
            code: "NOT_FOUND"
...